% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_to_igr.R
\name{ig_to_igr}
\alias{ig_to_igr}
\title{Convert Irish Grid coordinates to Irish Grid References}
\usage{
ig_to_igr(x, digits = 3)
}
\arguments{
\item{x}{a list of two character vectors containing Irish Grid eastings and northings respectively.}

\item{digits}{the resolution of grid reference required, between 0 (100km) and 6 (1m).}
}
\value{
A character vector of Irish Grid References.
}
\description{
Convert Irish Grid coordinates to Irish Grid References
}
\examples{
ig_to_igr(list(x = 0, y = 0))
ig_to_igr(list(x = c(0, 400000), y = c(0, 40000)))
}
