% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_to_igr.R
\name{ig_to_igr}
\alias{ig_to_igr}
\title{Convert Irish Grid coordinates to Irish Grid References}
\usage{
ig_to_igr(x, digits = 3, sep = "")
}
\arguments{
\item{x}{a list of two character vectors containing Irish Grid eastings and northings respectively.}

\item{digits}{an integer, the number of digits for both easting and northing, between 0 (100 km resolution) and 5 (1 m resolution).}

\item{sep}{a character to separate the 100km grid letter, easting, and northing.}
}
\value{
A character vector of Irish Grid References.
}
\description{
Convert Irish Grid coordinates to Irish Grid References
}
\examples{
ig_to_igr(list(x = 0, y = 0))
ig_to_igr(list(x = c(0, 400000), y = c(0, 40000)))
ig_to_igr(list(x = c(0, 400000), y = c(0, 40000)), sep = " ")
ig_to_igr(list(x = c(0, 400000), y = c(0, 40000)), sep = " ")
}
