% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_to_igr.R
\name{ig_to_igr}
\alias{ig_to_igr}
\title{Convert Irish Grid coordinates to Irish grid references}
\usage{
ig_to_igr(x, digits = 3, sep = "")
}
\arguments{
\item{x}{a matrix containing Irish Grid eastings and northings in the first
and second columns respectively.}

\item{digits}{an integer, the number of digits for both easting and northing
in the Irish grid references. 0 is equivalent to a resolution of 100 km, 1:
10 km, 2: 1 km, 3: 100 m, 4: 10 m, and 5: 1 m.}

\item{sep}{a character to separate the 100 km grid letter, easting, and
northing.}
}
\value{
A character vector of Irish grid references.
}
\description{
\code{ig_to_igr()} returns the Irish grid references for valid Irish Grid
(EPSG:29903) coordinates, otherwise NA.
}
\examples{
# A matrix of three Irish Grid coordinates
m <- matrix(c(0, 412300, 0, 98700, 456000, 0), byrow = TRUE, ncol = 2)

m

# Convert to Irish grid references
ig_to_igr(m)

# Insert a space between the 100 km grid letter, easting, and northing
ig_to_igr(m, sep = " ")

# Convert into Irish grid references with 1 km resolution
ig_to_igr(m, digits = 2)
}
