% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igr-package.R
\docType{package}
\name{igr-package}
\alias{igr}
\alias{igr-package}
\title{igr: Irish Grid Reference Utilities}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Convert between Irish grid references and Irish Grid coordinates. Irish grid references can also be converted to or from an 'sf' object in any coordinate reference system. Precisions from 1 m to 100 km are supported, as are datasets with mixed precision. Conversion to 'sf' polygons is precision-aware.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/digitalnature-ie/igr}
  \item \url{https://digitalnature-ie.github.io/igr/}
  \item Report bugs at \url{https://github.com/digitalnature-ie/igr/issues}
}

}
\author{
\strong{Maintainer}: John Kennedy \email{john@digitalnature.ie} (\href{https://orcid.org/0000-0002-1128-987X}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Digital Nature [funder]
}

}
\keyword{internal}
