% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igr_to_ig.R
\name{igr_to_ig}
\alias{igr_to_ig}
\title{Convert Irish Grid References to Irish Grid coordinates}
\usage{
igr_to_ig(igr, coords = c("x", "y"))
}
\arguments{
\item{igr}{A character vector of Irish Grid references.}

\item{coords}{A character vector of column names to contain the easting and northing.}
}
\value{
A list of Irish Grid coordinates.
}
\description{
Convert Irish Grid References to Irish Grid coordinates
}
\examples{
igr_to_ig("A00")
igr_to_ig(c("N8090", "D1234588800"))
}
