% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igr_to_ig.R
\name{igr_to_ig}
\alias{igr_to_ig}
\title{Convert Irish grid references to Irish Grid coordinates}
\usage{
igr_to_ig(x, coords = c("x", "y"), res = NULL)
}
\arguments{
\item{x}{A character vector of Irish grid references. Irish grid references
consist of a letter (other than I) optionally followed by both an easting
and northing. The easting and northing must be of the same length of
between 1 and 5 digits. There may be whitespace between the letter, easting
and northing.}

\item{coords}{A character vector of column names for easting and northing.}

\item{res}{A character vector: the column name for resolution of original
grid reference in metres, if required.}
}
\value{
A list containing Irish Grid x and y coordinates and (optionally)
the resolution in metres of each Irish grid reference. Invalid Irish grid
references return NA.
}
\description{
\code{igr_to_ig()} converts Irish grid references into a list of Irish Grid
(EPSG:29903) x and y coordinates and (optionally) grid reference resolutions
in metres.
}
\examples{
# Convert an Irish grid reference to Irish Grid coordinate
igr_to_ig("A00")

# Convert a vector of Irish grid references to Irish Grid coordinates
igr_to_ig(c("N8090", "D1234588800"))

# Use specific column names
igr_to_ig(c("N8090", "D1234588800"), coords = c("e", "n"))

# Also return the resolution in metres of each grid reference
igr_to_ig(c("N8090", "D1234588800"), res = "res")
}
