% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igr_to_ig.R
\name{igr_to_ig}
\alias{igr_to_ig}
\title{Title}
\usage{
igr_to_ig(igr)
}
\arguments{
\item{igr}{A character vector of Irish Grid References.}
}
\value{
A something of Irish Grid coordinates.
}
\description{
Title
}
\examples{
igr_to_ig("A00")
igr_to_ig(c("N8090", "D1234588800"))
}
