% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_igr_as_sf.R
\name{st_igr_as_sf}
\alias{st_igr_as_sf}
\title{Convert data frame containing Irish Grid References to an sf object}
\usage{
st_igr_as_sf(
  x,
  igrefs,
  crs = 29903,
  remove = FALSE,
  add_coords = FALSE,
  coords = c("x", "y")
)
}
\arguments{
\item{x}{object to be converted into an object class sf. Must not contain columns specified in coords.}

\item{igrefs}{name or number of the character column holding Irish Grid References.}

\item{crs}{coordinate reference system to be assigned; object of class crs.}

\item{remove}{logical; remove Irish Grid References column?}

\item{add_coords}{logical; add Irish Grid coordinate columns?}

\item{coords}{A character vector of column names to contain the Irish Grid easting and northing.}
}
\value{
An sf object.
}
\description{
Convert data frame containing Irish Grid References to an sf object
}
\examples{
x <- data.frame(igr = c("A00", "N8000"))
st_igr_as_sf(x, "igr")
st_igr_as_sf(x, "igr", crs = 4326)
}
