% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_igr_as_sf.R
\name{st_igr_as_sf}
\alias{st_igr_as_sf}
\title{Convert data frame containing Irish grid references to an sf object}
\usage{
st_igr_as_sf(
  x,
  igrefs = "igr",
  crs = 29903,
  remove = FALSE,
  add_coords = FALSE,
  coords = c("x", "y"),
  precision = NULL,
  polygons = FALSE
)
}
\arguments{
\item{x}{object containing column of Irish grid references. Must not contain
columns specified in \code{coords}, or invalid Irish grid references.}

\item{igrefs}{string (column name) or number (column index) of the character
column holding Irish grid references.}

\item{crs}{object of class crs containing the coordinate reference system to
be assigned. See \code{\link[sf:st_crs]{sf::st_crs()}}.}

\item{remove}{a logical indicating if the column containing the Irish grid
references should be removed.}

\item{add_coords}{a logical indicating if Irish Grid coordinate columns
should be added.}

\item{coords}{A character vector of column names for easting and northing.}

\item{precision}{A single character vector: the column name for precision of
original grid reference in metres, if required.}

\item{polygons}{a logical indicating if polygon features spanning the extent
of each grid reference should be returned, rather than point features.}
}
\value{
An sf object containing point or polygon features for each Irish grid
reference in x.
}
\description{
\code{st_igr_as_sf()} transforms Irish grid references into an sf object
containing point or polygon features. Points are the south west corner of the
grid references. Polygons are Irish Grid squares covering the full extent of
the grid references, each square size depending on the precision of each grid
reference. The Irish Grid (EPSG:29903) x and y coordinates and grid reference
precision in metres can be appended.
}
\examples{
# A data.frame containing two Irish grid references
x <- data.frame(igr = c("A00", "N8000"))

# Convert a data.frame of Irish grid references to an sf object in the
# Irish Grid coordinate reference system
st_igr_as_sf(x, "igr")

# Convert to an sf object in the WGS 84 coordinate reference system
st_igr_as_sf(x, "igr", crs = 4326)

# Include the Irish Grid coordinates and precision in the sf object
st_igr_as_sf(x, "igr", add_coords = TRUE, precision = "prec")

# Convert into polygons rather than points
st_igr_as_sf(x, "igr", polygons = TRUE)

}
