% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_irishgridrefs.R
\name{st_irishgridrefs}
\alias{st_irishgridrefs}
\title{Generate Irish grid references from sf point data}
\usage{
st_irishgridrefs(x, digits = 3, sep = "")
}
\arguments{
\item{x}{an sf object containing geometries of type POINT.}

\item{digits}{an integer, the number of digits for both easting and northing
in the Irish grid references. 0 is equivalent to a resolution of 100 km, 1:
10 km, 2: 1 km, 3: 100 m, 4: 10 m, and 5: 1 m.}

\item{sep}{a character to separate the 100 km grid letter, easting, and
northing.}
}
\value{
a character vector of Irish grid references.
}
\description{
\code{st_irishgridrefs()} returns the Irish grid references for point geometries
in an sf object that are located within the Irish Grid (EPSG:29903), and NA
otherwise.
}
\examples{
# An sf object containing point data
x_sf <- sf::st_as_sf(data.frame(x = c(0, 490000), y = c(400000, 0)),
  crs = 29903,
  coords = c("x", "y")
)

# Convert to Irish grid references
st_irishgridrefs(x_sf)

# Convert to Irish grid references at 100 km resolution
st_irishgridrefs(x_sf, digits = 0)

# Convert to Irish grid references at 1 m resolution
st_irishgridrefs(x_sf, digits = 5)

# Insert a space between the 100 km grid letter, easting, and northing
st_irishgridrefs(x_sf, sep = " ")
}
