% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_irishgridrefs.R
\name{st_irishgridrefs}
\alias{st_irishgridrefs}
\title{Generate Irish Grid References from sf point data}
\usage{
st_irishgridrefs(x, digits = 3, sep = "")
}
\arguments{
\item{x}{an sf object containing point data within the Irish Grid bounding box.}

\item{digits}{an integer, the number of digits for both easting and northing, between 0 (100 km resolution) and 5 (1 m resolution).}

\item{sep}{a character to separate the 100km grid letter, easting, and northing.}
}
\value{
a character vector of Irish Grid References.
}
\description{
Generate Irish Grid References from sf point data
}
\examples{
x_sf <- sf::st_as_sf(data.frame(x = c(0, 490000), y = c(400000, 0)),
  crs = 29903,
  coords = c("x", "y")
)
st_irishgridrefs(x_sf)
st_irishgridrefs(x_sf, digits = 0)
st_irishgridrefs(x_sf, digits = 5)
}
